agent-code/requirements.txt
azure-storage-queue
azure-identity